refactor: split lms custom course and shortform modules
